Rename api/actions/utils to createApiAction
The createApiAction function is the main/only thing needed to create
api actions. Other functions are exported for testing purposes. For
that reason it makes sense to rename the file.
Refs #10.
